+++ R/iso-year-week-day.R
@@ -869,6 +869,73 @@ calendar_widen.clock_iso_year_week_day <- function(x, precision) {
 
 # ------------------------------------------------------------------------------
 
+#' Boundaries: iso-year-week-day
+#'
+#' This is an iso-year-week-day method for the [calendar_start()] and
+#' [calendar_end()] generics. They adjust components of a calendar to the
+#' start or end of a specified `precision`.
+#'
+#' @inheritParams iso-year-week-day-group
+#'
+#' @return `x` at the same precision, but with some components altered to be
+#'   at the boundary value.
+#'
+#' @name iso-year-week-day-boundary
+#'
+#' @examples
+#' x <- iso_year_week_day(2019:2020, 5, 6, 10)
+#' x
+#'
+#' # Compute the last moment of the last iso week of the year
+#' calendar_end(x, "year")
+#'
+#' # Compare that to just setting the week to `"last"`,
+#' # which doesn't affect the other components
+#' set_week(x, "last")
+NULL
+
+#' @rdname iso-year-week-day-boundary
+#' @export
+calendar_start.clock_iso_year_week_day <- function(x, precision) {
+  x_precision <- calendar_precision_attribute(x)
+  precision <- validate_precision_string(precision)
+
+  calendar_start_end_checks(x, x_precision, precision, "start")
+
+  if (precision <= PRECISION_YEAR && x_precision > PRECISION_YEAR) {
+    x <- set_week(x, 1L)
+  }
+  if (precision <= PRECISION_WEEK && x_precision > PRECISION_WEEK) {
+    x <- set_day(x, 1L)
+  }
+
+  x <- calendar_start_time(x, x_precision, precision)
+
+  x
+}
+
+#' @rdname iso-year-week-day-boundary
+#' @export
+calendar_end.clock_iso_year_week_day <- function(x, precision) {
+  x_precision <- calendar_precision_attribute(x)
+  precision <- validate_precision_string(precision)
+
+  calendar_start_end_checks(x, x_precision, precision, "end")
+
+  if (precision <= PRECISION_YEAR && x_precision > PRECISION_YEAR) {
+    x <- set_week(x, "last")
+  }
+  if (precision <= PRECISION_WEEK && x_precision > PRECISION_WEEK) {
+    x <- set_day(x, 7L)
+  }
+
+  x <- calendar_end_time(x, x_precision, precision)
+
+  x
+}
+
+# ------------------------------------------------------------------------------
+
 #' Sequences: iso-year-week-day
 #'
 #' @description