+++ tests/testthat/test-iso-year-week-day.R
@@ -150,6 +150,48 @@ test_that("can widen to day", {
 })
 
 # ------------------------------------------------------------------------------
+# calendar_start()
+
+test_that("can compute year start", {
+  x <- iso_year_week_day(2019)
+  expect_identical(calendar_start(x, "year"), x)
+
+  x <- iso_year_week_day(2019, 2, 2, 2, 2, 2, 2, subsecond_precision = "millisecond")
+  expect <- iso_year_week_day(2019, 1, 1, 0, 0, 0, 0, subsecond_precision = "millisecond")
+  expect_identical(calendar_start(x, "year"), expect)
+})
+
+test_that("can compute week start", {
+  x <- iso_year_week_day(2019, 2)
+  expect_identical(calendar_start(x, "week"), x)
+
+  x <- iso_year_week_day(2019, 2, 2, 2, 2, 2, 2, subsecond_precision = "millisecond")
+  expect <- iso_year_week_day(2019, 2, 1, 0, 0, 0, 0, subsecond_precision = "millisecond")
+  expect_identical(calendar_start(x, "week"), expect)
+})
+
+# ------------------------------------------------------------------------------
+# calendar_end()
+
+test_that("can compute year end", {
+  x <- iso_year_week_day(2019)
+  expect_identical(calendar_end(x, "year"), x)
+
+  x <- iso_year_week_day(2019:2020, 2, 2, 2, 2, 2, 2, subsecond_precision = "millisecond")
+  expect <- iso_year_week_day(2019:2020, 52:53, 7, 23, 59, 59, 999L, subsecond_precision = "millisecond")
+  expect_identical(calendar_end(x, "year"), expect)
+})
+
+test_that("can compute week end", {
+  x <- iso_year_week_day(2019, 2)
+  expect_identical(calendar_end(x, "week"), x)
+
+  x <- iso_year_week_day(2019, 2, 2, 2, 2, 2, 2, subsecond_precision = "millisecond")
+  expect <- iso_year_week_day(2019, 2, 7, 23, 59, 59, 999L, subsecond_precision = "millisecond")
+  expect_identical(calendar_end(x, "week"), expect)
+})
+
+# ------------------------------------------------------------------------------
 # seq()
 
 test_that("only year precision is allowed", {